Hybrid_rag_Fastapi_Streamlit/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pdfplumber
python-docx
langchain
langchain-openai
langchain-community
streamlit
requests
python-multipart